fix: clean up exports, fix layer bugs, add test coverage
- Remove deprecated Medusa and MTP components from llminfra/inference/__init__.py exports
- Fix attention layer validations: add checks for mask sequence length mismatch and float masks, with corresponding test cases
- Correct LightningAttention scale factor to use feature_dim instead of head_dim, fix mask handling for 3D/4D inputs
- Fix RotaryPositionEmbedding to compute frequencies in float32 to avoid precision loss on long sequences
- Fix ALiBi positional encoding to use local tensor references instead of stale buffer copies
- Fix BlockSparseIndexer to return proper zero tensors for empty sequence lengths
- Fix PagedAttention einsum order and causal mask dimensions, add empty sequence handling
- Fix TieredKVCache to store detached clones of key/value tensors instead of raw detached views
- Fix ManifoldConstrainedHyperConnection to prevent logit overflow during sinkhorn normalization, add stream scaling/biases and test cases
- Fix transformer block norm initialization for deepnorm and post-norm styles
- Fix CrossAttention forward method override with proper type ignore and documentation about deviating from BaseAttention contract
- Fix flash attention v2/v3/v4 debug tracing to only record traces when keep_debug_state is enabled
- Fix positional encoding factory to reject extra kwargs for 2D positional embeddings
- Add input validation and documentation for EagleSpeculator, MTPDecoder, and NGramSpeculator
- Fix CausalLMModel default GQA group calculation to support odd numbers of attention heads
- Fix multimodal model pooling to respect attention masks, add validation for image grid token counts
- Add new tests for edge cases including empty sequences, mixed dtypes in flash attention, invalid masks, and hyperconnection behavior

+++ llminfra/positional/rotary.py
@@ -64,10 +64,12 @@ class RotaryPositionEmbedding(BasePositionalEncoding):
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """Rotate ``x`` by position indices derived from its sequence length."""
         seq_len = x.size(-2)
-        positions = torch.arange(seq_len, device=x.device, dtype=x.dtype)
-        freqs = torch.einsum("s,f->sf", positions, self.inv_freq.to(x.dtype))
-        cos = torch.cos(freqs)
-        sin = torch.sin(freqs)
+        # Compute frequencies in float32 so half-precision inputs do not
+        # lose positional accuracy on long sequences.
+        positions = torch.arange(seq_len, device=x.device, dtype=torch.float32)
+        freqs = torch.einsum("s,f->sf", positions, self.inv_freq.to(torch.float32))
+        cos = torch.cos(freqs).to(x.dtype)
+        sin = torch.sin(freqs).to(x.dtype)
         return apply_rotary_pos_emb(x, cos, sin)
 
     def extra_repr(self) -> str:
